payment should work idk
