+++ src/models/Post.ts
@@ -12,9 +12,6 @@ interface IPost extends Document {
   tags: string;
   others: string;
   date: Date;
-  paymentSecret: string;
-  paymentStatus: string;
-  paymentId: string;
   createdAt: Date;
   updatedAt: Date;
 }
@@ -102,16 +99,6 @@ const PostSchema = new Schema(
     date: {
       type: Date,
       required: true,
-    },
-    paymentSecret: {
-      type: String,
-    },
-    paymentStatus: {
-      type: String,
-      enum: ['pending', 'processing', 'succeeded'],
-    },
-    paymentId: {
-      type: String,
     },
   },
   { timestamps: true },
